Extending functionality for JTACYearView and Delegates #1310
Added support for scrolling year view to a specified date and a new delegate method to allow for detection of user selection of a given month view. Changed visibility of the extension to provide default protocol implementations so apps using library will benefit. Extended sample app to show new functionality in year view